correct ClassCastException: java.lang.Integer cannot be cast to java.lang.Long reported by Deepak Dixit. Problem raise through http://demo-trunk-ofbiz.apache.org/content/control/stream?contentId= where content is associate to ElectronicText data resource
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1628167 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/content/src/org/ofbiz/content/data/DataResourceWorker.java
@@ -1017,7 +1017,7 @@ public class DataResourceWorker  implements org.ofbiz.widget.DataResourceWorkerI
             }
 
             byte[] bytes = text.getBytes();
-            return UtilMisc.toMap("stream", new ByteArrayInputStream(bytes), "length", Integer.valueOf(bytes.length));
+            return UtilMisc.toMap("stream", new ByteArrayInputStream(bytes), "length", Long.valueOf(bytes.length));
 
         // object (binary) data
         } else if (dataResourceTypeId.endsWith("_OBJECT")) {
